aider: Set profile picture in menu and profile page dynamically based on user's profile data.
# Aider chat conversation:

ASSISTANT: To make the profile picture in the menu the same as the one in the profile page, we need to perform the following steps:

1. Modify `src/lib/components/menu.svelte` to use a dynamic source for the profile picture instead of the static `/profile.png`.
2. Modify `src/routes/profile/+page.svelte` to set the profile picture source dynamically, based on the user's profile data.

Let's start with the `src/lib/components/menu.svelte` file:

src/lib/components/menu.svelte
```svelte
<<<<<<< SEARCH
    <a href="/profile" class="h-11 w-11"><img src="/profile.png" alt="profile" class="rounded-full"/></a>
=======
    <a href="/profile" class="h-11 w-11"><img src="{user.profilePic || '/default-profile.png'}" alt="profile" class="rounded-full"/></a>
>>>>>>> REPLACE
```

In the above block, we assume that there is a `user` store or object that contains a `profilePic` property which holds the URL to the user's profile picture. If the `profilePic` is not set, it falls back to `'/default-profile.png'`.

Next, let's ensure that the profile page is setting the `user.profilePic` correctly. For this, we need to modify the `src/routes/profile/+page.svelte` file:

src/routes/profile/+page.svelte
```svelte
<<<<<<< SEARCH
<center class="pb-5 pt-10"><img width="225" class="rounded-full" src="profile.png" alt="profile image"></center>
=======
<script>
  import { user } from '$lib/stores'; // Import the user store
</script>

<center class="pb-5 pt-10"><img width="225" class="rounded-full" src="{user.profilePic || '/default-profile.png'}" alt="profile image"></center>
>>>>>>> REPLACE
```

Again, we assume that there is a `user` store that contains the user's profile data, including the `profilePic` property. If the `profilePic` is not set, it falls back to `'/default-profile.png'`.

Please note that the actual implementation of the `user` store and how the profile picture URL is set and updated will depend on your application's logic. You may need to adjust the code to fit your specific use case. If you need help with the implementation of the `user` store or the logic to set the profile picture URL, please provide more details or the relevant code.