aider: Fix incorrect usage of `user` store in `src/routes/setup/profile/+page.svelte`.
# Aider chat conversation:

USER: same error. fix it.

ASSISTANT: It seems there might be an issue with the way the `user` store is being used in the `src/routes/setup/profile/+page.svelte` file. The `user` store is a Svelte writable store, and to access its value within a component, we need to use the `$` syntax to subscribe to it.

Let's correct the usage of the `user` store in the `uploadProfilePicture` function by subscribing to the store to get the user's `uid`.

Here is the *SEARCH/REPLACE block* for the necessary changes:

src/routes/setup/profile/+page.svelte
```javascript
<<<<<<< SEARCH
    const storageRef = ref(storage, 'profile-pictures/' + user.uid + '.png');
=======
    const storageRef = ref(storage, 'profile-pictures/' + $user.uid + '.png');
>>>>>>> REPLACE
```

This change ensures that we are correctly accessing the `uid` property of the `user` store's current value when creating the storage reference for the profile picture. If there are other places in the code where the `user` store's value is accessed without the `$` syntax, similar changes will need to be made.